Use helper `saved_for_the_first_time` method in document_spec
This commit reduces duplication, and groups some methods in a dedicated
helper module.
